Collect-Fees/requirements.txt
requests>=2.28.0
python-decouple>=3.6